javadoc and minor code cleanup
